Из config.ts вынесены импорты, провоцирующие бесконечную цепь импортов
