+++ source/app/App.ts
@@ -1,7 +1,7 @@
 import "./ui/index.scss";
 import { Router } from "@/shared/Router/Router.ts";
 import { UserStorage } from "@/entities/User";
-import { routes, strictRoutes, defaultAuthRoutes } from "./config.ts";
+import { defaultAuthRoutes, routes, strictRoutes } from "@/shared/Router/Routes.ts";
 /**
  * Class provides class App, the initial class
  */
